Moved reducers, actions and constants into respective general files.
