+++ src/app/config/constants.js
@@ -85,3 +85,5 @@ export const ADD_GROUPS_TO_USER_PROGRESS = "ADD_GROUPS_TO_USER_PROGRESS";
 export const SIGN_OUT_ALL_USERS_FAILURE = "SIGN_OUT_ALL_USERS_FAILURE";
 export const SIGN_OUT_ALL_USERS_SUCCESS = "SIGN_OUT_ALL_USERS_SUCCESS";
 export const SIGN_OUT_ALL_USERS_PROGRESS = "SIGN_OUT_ALL_USERS_PROGRESS";
+
+export const SET_PREVIEW_LANGUAGE = "SET_PREVIEW_LANGUAGE";
