+++ src/app/config/initialState.js
@@ -62,4 +62,5 @@ export const initialState = {
     interactives: {
         all: [],
     },
+    previewLanguage: "en",
 };
