--- src/app/config/language/languageActions.js
@@ -1,8 +1,0 @@
-import * as types from "./languageConstants";
-
-export const setPreviewLanguage = language => {
-    return {
-        type: types.SET_PREVIEW_LANGUAGE,
-        language,
-    };
-};
