--- src/app/config/language/languageConstants.js
@@ -1,1 +1,0 @@
-export const SET_PREVIEW_LANGUAGE = "SET_PREVIEW_LANGUAGE";
